Introduce flag to warn on passing pointer to function
Adds a new flag `-disallowFuncs` which causes exportloopref to
emit a diagnostic if a reference to the loop variable is passed
to any function other than some explicitly allowed functions
(namely, functions in the `fmt` package, and the `print` and
`println` built-ins).

Also adds a test for this flag and its behavior, and updates
documentation to point users to the flag.